Create Avery Holloman's Portfolio
# About Me

Hello, I'm Avery M. Holloman, a passionate data enthusiast based in Lakeland, FL. I'm a recent graduate with a Bachelor of Science in Data Science & Analytics from the University of South Florida. My academic journey has equipped me with a strong foundation in data analysis, programming, and a wide array of specialized skills in the field. I'm actively seeking a full-time entry-level position in the data science, data analytics, or data engineering domains to apply and further develop my expertise.

## Education

My academic journey was a rewarding one, culminating in my Bachelor's degree with a major in Information Studies and a concentration in Data Science and Analytics. Throughout my studies, I tackled complex subjects such as advanced statistics, predictive analytics, and programming with a focus on R and Python. My coursework provided me with the skills and knowledge to work on a variety of data-related projects, from data visualization to predictive modeling and text analysis.

## Technical Skills

In the world of data, versatility is key, and I've honed my skills in various areas:

- **Programming Languages:** I'm proficient in C++, R, and Python and have experience working with data analysis libraries such as Matplotlib, R Studio, and Tableau.

- **Machine Learning:** I've dived into supervised learning, regression, and clustering analysis, utilizing tools like caret in R to build and evaluate predictive models.

- **Data Visualization:** With a strong foundation in ggplot2 and data visualization techniques, I can transform raw data into insightful visualizations.

- **Text Mining:** I've explored text analysis in R and worked with bag-of-words models to extract valuable insights from unstructured text data.

## Experience

I bring practical experience from my roles in the tech industry:

- **Omnicell (June 2023 - September 2023):** As a Technical Support Advisor, I understood Omnicell solutions, collaborated with cross-functional teams, and executed technical solutions while adhering to software standards and Agile principles.

- **Liberty Dental Plan (April 2017 - June 2023):** I served as a Technical Support Engineer, resolving customer service issues, generating reports, and ensuring the smooth operation of voice, data, and video services.

- **Verizon Business (June 2007 - January 2017):** In my role as a DSL Billing and Technical Support Team Supervisor, I managed a team of specialists and improved customer satisfaction through process enhancements and training sessions.

## Let's Collaborate

I'm excited to be a part of the data science community on GitHub. I'm constantly exploring new projects and opportunities, and I'm open to collaboration and learning from fellow data enthusiasts. Whether it's coding, data analysis, or machine learning, I'm ready to dive into exciting projects and expand my knowledge.

Feel free to connect with me on [LinkedIn](https://www.linkedin.com/in/averyhollomanusf2022) for more in-depth discussions, or explore my code repositories here on GitHub. Let's innovate and explore the endless possibilities of data together!

Looking forward to connecting with you and exploring the data-driven world.